+++ bioimageio/core/contrib/__init__.py
@@ -1,0 +1,1 @@
+from .contrib_a import *
