+++ bioimageio/core/contrib/contrib_a/__init__.py
@@ -1,0 +1,4 @@
+try:
+    from .local import *
+except ImportError:
+    from .remote import *
