+++ bioimageio/core/contrib/contrib_a/local.py
@@ -1,0 +1,2 @@
+from ._demo import hello
+from ._inference import run_model_inference_with_dask
