+++ bioimageio/core/operators/__init__.py
@@ -1,4 +1,4 @@
 from ._assert import assert_shape
 from ._generate import generate_random_uniform_tensor
-from ._run import run_model_inference, run_model_inference_without_tiling, run_workflow
+from ._run import WorkflowState, run_model_inference_without_tiling, run_workflow, run_workflow_test
 from ._various import binarize, load_tensors, log, select_outputs
