+++ dev/environment-base.yaml
@@ -18,3 +18,4 @@ dependencies:
   - tifffile <=2022.4.8  # pin fixes Syntax error; see https://github.com/bioimage-io/core-bioimage-io-python/pull/259
   - pip:
     - keras==1.2.2
+    - imjoy-rpc
